Move business-level exceptions to their own folder
