+++ src/main/java/ca/utoronto/fitbook/adapter/persistence/GenericRepository.java
@@ -1,6 +1,6 @@
 package ca.utoronto.fitbook.adapter.persistence;
 
-import ca.utoronto.fitbook.application.port.in.EntityNotFoundException;
+import ca.utoronto.fitbook.application.exceptions.EntityNotFoundException;
 
 public interface GenericRepository<T>
 {
