+++ src/main/java/ca/utoronto/fitbook/adapter/persistence/localmemory/UserLocalMemoryRepository.java
@@ -1,7 +1,7 @@
 package ca.utoronto.fitbook.adapter.persistence.localmemory;
 
 import ca.utoronto.fitbook.adapter.persistence.GenericRepository;
-import ca.utoronto.fitbook.application.port.in.EntityNotFoundException;
+import ca.utoronto.fitbook.application.exceptions.EntityNotFoundException;
 import ca.utoronto.fitbook.application.port.in.LoadUserPort;
 import ca.utoronto.fitbook.entity.User;
