+++ src/main/java/ca/utoronto/fitbook/application/port/in/LoadUserByNamePort.java
@@ -1,5 +1,7 @@
 package ca.utoronto.fitbook.application.port.in;
 
+import ca.utoronto.fitbook.application.exceptions.UsernameCollisionException;
+import ca.utoronto.fitbook.application.exceptions.UsernameNotFoundException;
 import ca.utoronto.fitbook.entity.User;
 
 public interface LoadUserByNamePort {
