+++ src/main/java/ca/utoronto/fitbook/application/port/in/LoadUserPort.java
@@ -1,5 +1,6 @@
 package ca.utoronto.fitbook.application.port.in;
 
+import ca.utoronto.fitbook.application.exceptions.EntityNotFoundException;
 import ca.utoronto.fitbook.entity.User;
 
 public interface LoadUserPort
